refactor: move user authentication dependencies to a separate file
